feat(payment): migrate to per-1M token pricing and add character generation routes
BREAKING CHANGE: Credit cost calculation migrated from fixed per-message to per-1M token pricing

**Payment System Migration:**
- Change credits_cost from INTEGER to DECIMAL(10,6) for precise financial tracking
- Migrate from fixed per-message cost to per-million-token pricing matching Google's model
- Use BigDecimal for all financial calculations to avoid floating-point rounding errors
- Implement dual cost tracking: base API cost (dollars, source of truth) + user-facing credits (with 20% markup)
- Credit value: 1 credit = $0.02
- Token pricing: gemini-2.5-flash ($0.30/$2.50 per 1M), gemini-2.5-pro ($1.25/$10.00 per 1M), gemini-2.5-flash-lite ($0.10/$0.40 per 1M)
- Customer pricing: 20% markup over base API costs ($0.36/$3.00, $1.50/$12.00, $0.12/$0.48 per 1M)

**Character Generation Routes:**
- Add REST endpoints for AI-powered character field generation
- POST /api/generation/character/field - Generate single character field
- POST /api/generation/character/complete - Generate complete character
- POST /api/generation/character/enhance - Enhance existing character
- POST /api/generation/lorebook/entry - Generate single lorebook entry
- POST /api/generation/lorebook/entries - Generate multiple lorebook entries
- POST /api/generation/scribe-assistant - General AI assistant endpoint
- All routes currently return stub implementations

**Structured Outputs:**
- Replace brittle JSON parsing with Gemini 2.5+ native structured outputs
- Implement type-safe character field generation with guaranteed JSON conformance
- Add style detection analysis with structured output validation

**Frontend Token Usage Display:**
- Add per-session token usage tracking in chat interface
- Display input/output token counts with per-1M token pricing
- Show session total cost in dollars with proper BigDecimal handling
- Fix type safety issues in token-usage-display and chat components

**Database Schema:**
- Migration 2025-10-10-213932_change_credits_cost_to_decimal
- ALTER TABLE chat_messages: credits_cost INTEGER → DECIMAL(10,6)
- ALTER TABLE chat_sessions: total_credits_used INTEGER → DECIMAL(10,6)

Affected files:
- backend/migrations/2025-10-10-213932_change_credits_cost_to_decimal/
- backend/src/models/chats.rs (BigDecimal serialization)
- backend/src/routes/chat.rs (per-1M token pricing calculation)
- backend/src/routes/generation_routes.rs (NEW: character generation endpoints)
- backend/src/routes/mod.rs (register generation routes)
- backend/src/schema.rs (DECIMAL schema types)
- backend/src/services/chat/generation.rs (GenerationCost struct)
- backend/src/services/payment/credit_service.rs (BigDecimal cost tracking)
- backend/config/subscription_tiers.json (per-1M pricing configuration)
- backend/tests/*.rs (update all test fixtures to use BigDecimal::from(0))
- frontend/src/lib/components/chat.svelte (session cost display)
- frontend/src/lib/components/token-usage-display.svelte (per-1M pricing)
- frontend/src/lib/components/ai-assistant-dialog.svelte (style type safety)
- frontend/src/lib/types.ts (DescriptionStyle, token count types)

+++ backend/tests/chat_generate_stream_error_tests.rs
@@ -4,6 +4,7 @@ use axum::{
     body::Body,
     http::{Method, Request, StatusCode, header},
 };
+use bigdecimal::BigDecimal;
 use chrono::Utc;
 use diesel::prelude::*;
 use genai::chat::{ChatStreamEvent, StreamChunk};
@@ -153,7 +154,7 @@ async fn generate_chat_response_streaming_ai_error() {
                 total_completion_tokens: 0,
                 estimated_cost_cents: 0,
                 tokens_counted_at: chrono::Utc::now(),
-                total_credits_used: 0,
+                total_credits_used: BigDecimal::from(0),
                 prompt_template_id: "default".to_string(),
             };
             diesel::insert_into(chat_sessions_dsl::chat_sessions)
@@ -481,7 +482,7 @@ async fn generate_chat_response_streaming_initiation_error() {
                 total_completion_tokens: 0,
                 estimated_cost_cents: 0,
                 tokens_counted_at: chrono::Utc::now(),
-                total_credits_used: 0,
+                total_credits_used: BigDecimal::from(0),
                 prompt_template_id: "default".to_string(),
             };
             diesel::insert_into(chat_sessions_dsl::chat_sessions)
@@ -733,7 +734,7 @@ async fn generate_chat_response_streaming_error_before_content() {
                 total_completion_tokens: 0,
                 estimated_cost_cents: 0,
                 tokens_counted_at: chrono::Utc::now(),
-                total_credits_used: 0,
+                total_credits_used: BigDecimal::from(0),
                 prompt_template_id: "default".to_string(),
             };
             diesel::insert_into(chat_sessions_dsl::chat_sessions)
@@ -994,7 +995,7 @@ async fn generate_chat_response_streaming_genai_json_error() {
                 total_completion_tokens: 0,
                 estimated_cost_cents: 0,
                 tokens_counted_at: chrono::Utc::now(),
-                total_credits_used: 0,
+                total_credits_used: BigDecimal::from(0),
                 prompt_template_id: "default".to_string(),
             };
             diesel::insert_into(chat_sessions_dsl::chat_sessions)
