+++ backend/tests/chat_generate_stream_misc_tests.rs
@@ -4,6 +4,7 @@ use axum::{
     body::Body,
     http::{Method, Request, StatusCode, header},
 };
+use bigdecimal::BigDecimal;
 use chrono::Utc;
 use diesel::prelude::*;
 use genai::chat::{ChatStreamEvent, StreamChunk, StreamEnd};
@@ -156,7 +157,7 @@ async fn create_test_chat_session(
                 total_completion_tokens: 0,
                 estimated_cost_cents: 0,
                 tokens_counted_at: chrono::Utc::now(),
-                total_credits_used: 0,
+                total_credits_used: BigDecimal::from(0),
                 prompt_template_id: "default".to_string(),
             };
             diesel::insert_into(chat_sessions_dsl::chat_sessions)
@@ -777,7 +778,7 @@ async fn create_real_client_test_session(
                 total_completion_tokens: 0,
                 estimated_cost_cents: 0,
                 tokens_counted_at: chrono::Utc::now(),
-                total_credits_used: 0,
+                total_credits_used: BigDecimal::from(0),
                 prompt_template_id: "default".to_string(),
             };
             diesel::insert_into(chat_sessions_dsl::chat_sessions)
