+++ backend/tests/chat_suggested_actions_tests.rs
@@ -13,6 +13,7 @@ use uuid::Uuid;
 
 // Crate imports
 use anyhow::Context as _;
+use bigdecimal::BigDecimal;
 use scribe_backend::models::auth::LoginPayload;
 use scribe_backend::{
     PgPool,
@@ -123,7 +124,7 @@ async fn test_suggested_actions_success() -> anyhow::Result<()> {
         total_completion_tokens: 0,
         estimated_cost_cents: 0,
         tokens_counted_at: chrono::Utc::now(),
-        total_credits_used: 0,
+        total_credits_used: BigDecimal::from(0),
         prompt_template_id: "default".to_string(),
     };
 
@@ -358,7 +359,7 @@ async fn create_test_chat_session_for_suggested_actions(
         total_completion_tokens: 0,
         estimated_cost_cents: 0,
         tokens_counted_at: chrono::Utc::now(),
-        total_credits_used: 0,
+        total_credits_used: BigDecimal::from(0),
         prompt_template_id: "default".to_string(),
     };
